python-agent/requirements.txt
browser-use
langchain-openai
python-dotenv
playwright
reportlab
